dialogue-summarizer/requirements.txt
# Core dependencies
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
tiktoken>=0.5.0
